requirements.txt
absl-py==0.11.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
apex==0.1
appdirs==1.4.4
argon2-cffi==20.1.0
ascii-graph==1.5.1
ase==3.22.1
async-generator==1.10
async-timeout==4.0.1
attrs 
audioread==2.1.9
Babel==2.9.0
backcall==0.2.0
beautifulsoup4 
bleach==3.2.1
blessings==1.7
blis==0.4.1
boto3==1.16.25
botocore==1.19.25
brotlipy==0.7.0
cachetools==4.2.4
catalogue==1.0.0
certifi==2020.11.8
cffi 
chardet 
charset-normalizer==2.0.9
click==7.1.2
codecov==2.1.10
conda==4.9.2
conda-build==3.20.5
conda-package-handling 
coverage==5.3
cryptography 
cupy-cuda111==10.1.0
cxxfilt==0.2.2
cycler==0.10.0
cymem 
Cython==0.28.4
DataProperty==0.50.0
decorator==4.4.2
defusedxml==0.6.0
dgl==0.6.1
dgl-cu111==0.7.1
DLLogger 
docutils==0.16
dominate==2.6.0
entrypoints==0.3
fastrlock==0.8
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
glob2==0.7
google-auth==2.3.3
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
gpustat==0.6.0
graphsurgeon 
grpcio==1.33.2
h5py==3.1.0
html2text==2020.1.16
hypothesis==4.50.8
idna 
imageio==2.9.0
imagesize==1.2.0
importlib-metadata 
inflect==5.0.2
iniconfig==1.1.1
ipdb==0.13.4
ipykernel==5.3.4
ipython 
ipython_genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi 
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json5==0.9.5
jsonschema 
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyter-tensorboard==0.2.0
jupyterlab==1.2.14
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupytext==1.7.1
kiwisolver==1.3.1
libarchive-c==2.9
librosa==0.8.0
littleutils==0.2.2
llvmlite==0.34.0
lmdb==1.0.0
Markdown==3.3.3
markdown-it-py==0.5.6
MarkupSafe==1.1.1
maskrcnn-benchmark 
matplotlib==3.3.3
mbstrdecoder==1.0.0
mccabe==0.6.1
mistune==0.8.4
mlperf-compliance==0.0.10
mmcv==1.2.7
mmcv-full==1.2.7
mmdet==2.8.0
mmpycocotools==12.0.3
mock==4.0.2
msgfy==0.1.0
multidict==5.2.0
MultiScaleDeformableAttention==1.0
murmurhash==1.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.6.3
nltk==3.5
notebook==6.1.5
numba 
numpy 
nvidia-dali-cuda110==0.28.0
nvidia-ml-py3==7.352.0
nvidia-pyprof 
nvidia-tensorboard 
nvidia-tensorboard-plugin-dlprof 
oauthlib==3.1.1
ogb==1.3.1
onnx 
onnxruntime==1.5.2
opencv-contrib-python==4.5.5.64
opencv-python-headless==4.1.2.30
outdated==0.2.1
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.0
pathvalidate==2.3.0
pexpect 
pickleshare==0.7.5
Pillow==8.4.0
Pillow-SIMD 
pkginfo==1.6.1
plac 
pluggy==0.13.1
polygraphy==0.20.13
pooch==1.2.0
preshed==3.0.2
progressbar==2.5
prometheus-client==0.9.0
prompt-toolkit 
protobuf==3.14.0
psutil 
ptyprocess 
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycocotools 
pycodestyle==2.5.0
pycosat==0.6.3
pycparser 
pyDeprecate==0.3.0
pydot==1.4.1
pyflakes==2.1.1
Pygments 
pyinn==0.0.1
pynvml==8.0.4
pyOpenSSL 
pyparsing==2.4.7
pyrsistent 
PySocks 
pytablewriter==0.47.0
pytest==6.1.2
pytest-cov==2.10.1
pytest-pythonpath==0.7.3
python-dateutil==2.8.1
python-hostlist==1.21
python-louvain==0.15
python-nvd3==0.15.0
python-slugify==4.0.1
pytorch-lightning==1.3.1
pytorch-transformers==1.1.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
rdflib==6.0.2
rdkit-pypi==2021.9.2.1
regex==2020.11.13
requests 
requests-oauthlib==1.3.0
resampy==0.2.2
revtok 
rsa==4.8
ruamel_yaml==0.15.87
s3transfer==0.3.3
sacrebleu==1.2.10
sacremoses==0.0.35
scikit-image==0.15.0
scikit-learn==0.23.2
scipy 
Send2Trash==1.5.0
sentencepiece==0.1.94
six 
snowballstemmer==2.0.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
sox==1.4.1
spacy 
Sphinx==3.3.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
srsly==1.0.2
subword-nmt 
tabledata==1.1.3
tabulate==0.8.7
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorrt 
terminado==0.9.1
terminaltables==3.1.10
testpath==0.4.4
text-unidecode==1.3
thinc 
threadpoolctl==2.1.0
timm==0.3.2
toml==0.10.2
torch==1.7.1+cu110
torch-geometric==1.6.3
torch-scatter==2.0.6
torch-sparse==0.6.9
torchinfo==1.6.3
torchmetrics==0.6.1
torchsummary==1.5.1
torchtext==0.4.0
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.53.0
traitlets 
typepy==1.1.1
typing-extensions==3.7.4.3
uff 
Unidecode==1.1.1
urllib3 
wasabi 
wcwidth 
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.10.11
yacs==0.1.8
yapf==0.32.0
yarl==1.7.2
zipp 
